refactor(referentiels): extrait la partie drizzle de GetReferentielService dans un repositoy
pour être plus conforme avec nos conventions de code
